Cleaned up several lint warnings.
This involved renaming some .png to .jpg as it complained they were not
actually .png files. Need to address this in the future to allow for
transparency in the images.

Removed the colors.xml and used direct color values where previously
used. There were too many unused warnings in lint.