Converted statistics for Destroyer Game into a Fragment.
Fragment encapsulates the old behaviour for displaying the game
statistics but now allows for separation of the XML as well as
modularity. In addition, refactored some of the database fetching code
to be more robust.

Added in a (more) button for the Fragment to display a pop-up Dialog
that shows the entire list. This can be refined into a DialogFragment to
show more information if needed.

Removed length-checking code from AddEventFragment. Instead, used
maxLength within the related XML layout file to limit it.