Modified colors/appearance of DestroyerStatsFragment.
Modified AddEventFragment slightly; removed unnecessary listener as well
as class variables. Moved Intent code into createEvent() as this should
be paired together.

getActivity() should be called when a Context is needed in the rare case
that the owning Activity is changed (and thus the previous Context is no
longer valid).

Removed .orig XML files that make Eclipse angry; should be careful when
doing sync/merges as these are created and should be removed before
commit.