Changed how Games were locked/unlocked.
When a game is locked, a notice is displayed and statistics are hidden.

When a gem is unlocked, the notice is hidden and the statistics are
shown.

Moved generating values for statistics from onCreateView to onResume.

Removed commits to files in /gen as we these are rebuilt on a
per-machine basis.